[FREELDR] - Move cleaning the shared user data to after WinLdrSetProcessorContext, as we need the new page tables to be active - Fix indentation/coding style in registry.c - Fix a few 64 bit warnings - delete <arch>/ntsetup.c and rename <arch>/wlmemory.c to <arch>/winldr.c
svn path=/trunk/; revision=53526

+++ reactos/boot/freeldr/freeldr/fs/fat.c
@@ -650,7 +650,7 @@ static BOOLEAN FatXSearchDirectoryBufferForFile(PFAT_VOLUME_INFO Volume, PVOID D
 {
 	ULONG		EntryCount;
 	ULONG		CurrentEntry;
-	ULONG		FileNameLen;
+	SIZE_T		FileNameLen;
 	FATX_DIRENTRY	OurDirEntry;
 	PFATX_DIRENTRY	DirEntry = &OurDirEntry;
